% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountGenes.R
\name{CountGenes}
\alias{CountGenes}
\title{Count significantly changed genes from DESeq2 output}
\usage{
CountGenes(res_ordered)
}
\arguments{
\item{res_ordered}{sorted DESeq2 result table.}
}
\value{
A summary of significantly changed genes, based on padj<0.05 and padj<0.01 cutoffs.
}
\description{
Count significantly changed genes from DESeq2 output
}
