% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDE_FromRawCounts.R
\name{easyDE_FromRawCounts}
\alias{easyDE_FromRawCounts}
\title{Run DE analysis using raw counts}
\usage{
easyDE_FromRawCounts(
  count_matrix,
  LabelFile,
  ComparisonFile,
  createQuickomicsFiles = F,
  QuickomicsPrefix = NULL
)
}
\arguments{
\item{count_matrix}{data.frame of Raw count matrix}

\item{LabelFile}{File name of the Sample labels, with three columns: Data name, Data short name, condition}

\item{ComparisonFile}{DE comparison file with three columns: Condition_1 Condition_2 OutputPrefix}

\item{createQuickomicsFiles}{True or False, output Quickomics files}

\item{QuickomicsPrefix}{Prefix for Quickomics files}
}
\description{
Run DE analysis using raw counts
}
