% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDE_FromSalmon.R
\name{easyDE_FromSalmon}
\alias{easyDE_FromSalmon}
\title{Run DE analysis using salmon results}
\usage{
easyDE_FromSalmon(
  SampleInfo,
  uniqueMatchingFile,
  ComparisonFile,
  createQuickomicsFiles = F,
  QuickomicsPrefix = NULL
)
}
\arguments{
\item{SampleInfo}{File name of the Sample information text file.}

\item{uniqueMatchingFile}{Transcript to Gene unique matching file with two columns}

\item{ComparisonFile}{DE comparison file with three columns: Condition_1 Condition_2 OutputPrefix}

\item{createQuickomicsFiles}{True or False, output Quickomics files}

\item{QuickomicsPrefix}{Prefix for Quickomics files}
}
\description{
Run DE analysis using salmon results
}
