% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyEnrich_GO_GSEA.R
\name{easyEnrich_GO_GSEA}
\alias{easyEnrich_GO_GSEA}
\title{Run GO and GSEA analysis using DE results.}
\usage{
easyEnrich_GO_GSEA(
  ComparisonFile,
  OrganDatabase,
  GOpajd_cutoff = 0.05,
  GOlog2FC_cutoff = 1.5,
  ShowTermNum = 20,
  forcerun = F
)
}
\arguments{
\item{ComparisonFile}{DE comparison file with three columns: Condition_1 Condition_2 OutputPrefix. Comparison direction: Condition_2 v.s. Condition_1.}

\item{OrganDatabase}{Organism GO annotation database, for example, org.Mm.eg.db for mouse, and org.Hs.eg.db for human. Please install the required database packages before running this function.}

\item{GOpajd_cutoff}{pajd cutoff to define significant genes, default=0.05.}

\item{GOlog2FC_cutoff}{log2 fold change cutoff to define significant genes, default=1.5.}

\item{ShowTermNum}{Number of GOs to plot, default=20.}

\item{forcerun}{Force to run the analysis if DE gene number<10, default=False.}
}
\value{
For each comparison, three GO plots (BP, CC, and MF), a GSEA summary table, and a file containing Top 10 GSEA plots will be generated.
}
\description{
Run GO and GSEA analysis using DE results.
}
