% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featureCounts_Comparison}
\alias{featureCounts_Comparison}
\title{featureCounts_Comparison}
\format{
:
  featureCounts_Comparison: data.frame, example comparison table.
}
\usage{
featureCounts_Comparison
}
\description{
featureCounts_Comparison
}
\keyword{datasets}
