% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featureCounts_count_matrix}
\alias{featureCounts_count_matrix}
\title{List of pre-computed data}
\format{
:
  featureCounts_count_matrix: merged raw featureCounts table.
}
\usage{
featureCounts_count_matrix
}
\description{
featureCounts_count_matrix
}
\keyword{datasets}
