% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featureCounts_Label}
\alias{featureCounts_Label}
\title{featureCounts_Label}
\format{
:
  featureCounts_Label: data.frame, example label table.
}
\usage{
featureCounts_Label
}
\description{
featureCounts_Label
}
\keyword{datasets}
