% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeFeatureCounts.R
\name{MergeFeatureCounts}
\alias{MergeFeatureCounts}
\title{Merge featureCounts results to a table}
\usage{
MergeFeatureCounts(SampleInfoFile, OutputPrefix = "default", write_to_file = F)
}
\arguments{
\item{SampleInfoFile}{File name of the Sample information text file.}

\item{OutputPrefix}{Prefix of output files, default is "default".}

\item{write_to_file}{True of False, write to a file or not. If True, then use OutputPrefix as file prefix.}
}
\value{
Merged featureCounts table.
}
\description{
Merge featureCounts results to a table
}
