% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSalmon.R
\name{MergeSalmon}
\alias{MergeSalmon}
\title{Merge salmon results to a table}
\usage{
MergeSalmon(
  SampleInfoFile,
  uniqueMatchingFile,
  OutputPrefix = "default",
  write_to_file = F
)
}
\arguments{
\item{SampleInfoFile}{File name of the Sample information text file.}

\item{uniqueMatchingFile}{Transcript to Gene unique matching file with two columns.}

\item{OutputPrefix}{Prefix of output files, default is "default".}

\item{write_to_file}{True of False, write to a file or not. If True, then use OutputPrefix as file prefix.}
}
\value{
Merged salmon Counts and TPM table.
}
\description{
Merge salmon results to a table
}
