% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salmon_Comparison}
\alias{salmon_Comparison}
\title{salmon_Comparison}
\format{
:
  salmon_Comparison: data.frame, example comparison table.
}
\usage{
salmon_Comparison
}
\description{
salmon_Comparison
}
\keyword{datasets}
