% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salmon_uniqueMatching}
\alias{salmon_uniqueMatching}
\title{salmon_uniqueMatching}
\format{
:
  salmon_uniqueMatching: data.frame, example uniqueMatching file.
}
\usage{
salmon_uniqueMatching
}
\description{
salmon_uniqueMatching
}
\keyword{datasets}
