% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyDE_FromRawCounts.R
\name{easyDE_FromRawCounts}
\alias{easyDE_FromRawCounts}
\title{Run DE analysis using raw counts}
\usage{
easyDE_FromRawCounts(
  count_matrix,
  LabelFile,
  ComparisonFile,
  createQuickomicsFiles = F,
  QuickomicsPrefix = NULL
)
}
\arguments{
\item{count_matrix}{data.frame of raw count matrix. Can be generated by MergeFeatureCounts.}

\item{LabelFile}{File name of the Sample labels, with three columns: Data name, Data short name, condition.}

\item{ComparisonFile}{DE comparison file with three columns: Condition_1 Condition_2 OutputPrefix. Comparison direction: Condition_2 v.s. Condition_1.}

\item{createQuickomicsFiles}{True or False, output Quickomics files.}

\item{QuickomicsPrefix}{Prefix for Quickomics files.}
}
\value{
For each comparison, this function will generate a DE summary table (OutputPrefix.RawCounts.All.DEseq2GeneCountsNormcounts.txt) and multiple plots, including Distance plot, Heatmap, 2 PCA plots, Volcano plot, and MA plot.
}
\description{
Run DE analysis using raw counts
}
