% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyEnrich_KEGG.R
\name{easyEnrich_KEGG}
\alias{easyEnrich_KEGG}
\title{Run KEGG analysis using DE results.
Please run this function after running easyDE_FromRawCounts or easyDE_FromSalmon.
If you would like to prepare your own files, the input format should be DESeq2 output table.
The output results include KEGG enrichment (using significant DE genes) and GSEA results (using all ranked genes).
The output also includes a set of KEGG figures based on given KEGG pathway IDs.}
\usage{
easyEnrich_KEGG(
  ComparisonFile,
  OrganDatabase,
  KEGGSpeciesName,
  KEGGPathwayFile = NA,
  GOpajd_cutoff = 0.05,
  GOlog2FC_cutoff = 1.5,
  forcerun = F
)
}
\arguments{
\item{ComparisonFile}{DE comparison file with three columns: Condition_1 Condition_2 OutputPrefix.}

\item{OrganDatabase}{Organism GO annotation database, for example, 'org.Mm.eg.db' for mouse.}

\item{KEGGSpeciesName}{KEGG pathway species name, for example, 'mmu' for mouse.}

\item{KEGGPathwayFile}{A two-column file for KEGG pathways to plot: KEGG_id KEGG_full_name, default="NA"}

\item{GOpajd_cutoff}{pajd cutoff to define significant genes, default=0.05.}

\item{GOlog2FC_cutoff}{log2 fold change cutoff to define significant genes, default=0.05}

\item{forcerun}{Force to run the analysis if DE gene number<10, default=False.}
}
\description{
Run KEGG analysis using DE results.
Please run this function after running easyDE_FromRawCounts or easyDE_FromSalmon.
If you would like to prepare your own files, the input format should be DESeq2 output table.
The output results include KEGG enrichment (using significant DE genes) and GSEA results (using all ranked genes).
The output also includes a set of KEGG figures based on given KEGG pathway IDs.
}
